# Optimize ext/dom $wholeText